+++ src/dflow/context.py
@@ -1,8 +1,14 @@
+from .op_template import OPTemplate
+
+
 class Context(object):
     """
     Context
     """
-    def render(self, template):
+    def render(
+            self,
+            template : OPTemplate,
+    ) -> OPTemplate:
         """
         render original template and return a new template
         """